+++ app/server/controllers/errorController.js
@@ -8,6 +8,7 @@
  * Provides route actions to deal with errors.
  *
  * @class errorController
+ * @static
  */
 
 const errors = process.require('app/server/httpErrors.js');
@@ -18,6 +19,10 @@ const defaultController = process.require('app/server/controllers/defaultControl
  *
  * @method notFoundAction
  * @static
+ * @async
+ * @param {Request} request ExpressJS HTTP Request
+ * @param {Response} response ExpressJS HTTP Response
+ * @param {Function} next Function to defer execution to the next registered middleware
  */
 module.exports.notFoundAction = (request, response, next) => {
   next(errors.PATH_NOT_FOUND);
@@ -34,7 +39,11 @@ module.exports.notFoundAction = (request, response, next) => {
  *
  * @method errorAction
  * @static
+ * @async
  * @param {Object} error An error object with error code, HTTP code and error message
+ * @param {Request} request ExpressJS HTTP Request
+ * @param {Response} response ExpressJS HTTP Response
+ * @param {Function} next Function to defer execution to the next registered middleware
  */
 module.exports.errorAction = (error, request, response, next) => {
   if (!error)