+++ app/server/passport/cas/CAS2.js
@@ -9,7 +9,9 @@ const CAS = process.require('app/server/passport/cas/CAS.js');
 /**
  * Defines a cas client interfacing with cas protocol 2.0.
  *
- * @class CAS 2
+ * @class CAS2
+ * @extends CAS
+ * @constructor
  */
 class CAS2 extends CAS {