+++ server.js
@@ -3,26 +3,30 @@
 require('./processRequire.js');
 const path = require('path');
 const nopt = require('nopt');
-const openveoAPI = require('@openveo/api');
+const openVeoApi = require('@openveo/api');
 
 const conf = process.require('app/server/conf.js');
 const Server = process.require('app/server/Server.js');
-const configurationDirectoryPath = path.join(openveoAPI.fileSystem.getConfDir(), 'portal');
+const WebserviceClient = process.require('app/server/WebserviceClient.js');
+const configurationDirectoryPath = path.join(openVeoApi.fileSystem.getConfDir(), 'portal');
 const loggerConfPath = path.join(configurationDirectoryPath, 'loggerConf.json');
 const serverConfPath = path.join(configurationDirectoryPath, 'serverConf.json');
 const databaseConfPath = path.join(configurationDirectoryPath, 'databaseConf.json');
+const webservicesConfPath = path.join(configurationDirectoryPath, 'webservicesConf.json');
 const confPath = path.join(configurationDirectoryPath, 'conf.json');
 
 let loggerConf;
 let serverConf;
 let databaseConf;
+let webservicesConf;
 
 // Process list of arguments
 const knownProcessOptions = {
   serverConf: [String, null],
   databaseConf: [String, null],
   loggerConf: [String, null],
-  conf: [String, null]
+  conf: [String, null],
+  webservicesConf: [String, null]
 };
 
 // Parse process arguments
@@ -33,19 +37,23 @@ try {
   loggerConf = require(processOptions.loggerConf || loggerConfPath);
   serverConf = require(processOptions.serverConf || serverConfPath);
   databaseConf = require(processOptions.databaseConf || databaseConfPath);
+  webservicesConf = require(processOptions.webservicesConfPath || webservicesConfPath);
   conf.data = require(processOptions.conf || confPath);
 } catch (error) {
   throw new Error(`Invalid configuration file : ${error.message}`);
 }
 
 // Create logger
-process.logger = openveoAPI.logger.add('portal', loggerConf);
+process.logger = openVeoApi.logger.add('portal', loggerConf);
+
+// Create Web Service client
+WebserviceClient.create(webservicesConf);
 
 // Instanciate a Server
 const server = new Server(serverConf);
 
 // Get a Database instance
-const db = openveoAPI.Database.getDatabase(databaseConf);
+const db = openVeoApi.database.factory.get(databaseConf);
 
 // Establish connection to the database
 db.connect((error) => {
@@ -54,7 +62,6 @@ db.connect((error) => {
     throw new Error(error);
   }
 
-  openveoAPI.applicationStorage.setDatabase(db);
   server.onDatabaseAvailable(db);
   server.start();
 });